+++ src/protocols/ftps.rs
@@ -74,6 +74,9 @@ impl FileTransferClient for FtpsClient {
         port: u16,
         timeout: Duration,
         config: &ProtocolConfig,
+        _user: &str,
+        _password: Option<&str>,
+        _keyfile_path: Option<&str>,
     ) -> Result<Self, FtpError>
     where
         Self: Sized,